synapse_admin/requirements.txt
click
requests
pyyaml
tabulate
click-option-group>=0.5.2